redis: add hgetall and hdel commands
This patch adds support for 2 HASHes commands, and
internally introduce a different result type: hashes_result

I started with trying to reuse strings_result but it was
getting confusing so I decided to split it even if it's
a bit redundant. Please let me know if there is another way.

Signed-off-by: Etienne Adam <etienne.adam@gmail.com>